requirements.txt
# Uncomment to enable Azure Monitor OpenTelemetry
# Ref: aka.ms/functions-azure-monitor-python 
# azure-monitor-opentelemetry 

azure-functions==1.20.0
azure-storage-blob==12.8.1
azure-identity==1.12.0
numpy==1.26.4
pandas==2.2.2
scikit-learn==1.4.2
scipy==1.13.1
implicit
surprise